Move namespace for new classes (PR feedback) Fix perfview command-line (PR feedback)
